+++ pyLOM/NN/__init__.py
@@ -7,7 +7,7 @@
 # Last rev: 09/10/2024
 
 # Supress prints from tensorflow
-import os, torch, torch.nn as nn
+import os, torch
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'  
 from ..utils import MPI_RANK
 
@@ -16,17 +16,19 @@ DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
 if torch.cuda.is_available(): torch.cuda.set_device(int(MPI_RANK % torch.cuda.device_count())) 
 
 ## Flags to enchance performance when using torch compiled with cuda in the backend
-torch.backends.cudnn.enabled = True
+torch.backends.cudnn.enabled   = True
 torch.backends.cudnn.benchmark = True
 torch.set_float32_matmul_precision('high')
 torch.cuda.empty_cache()
 ALLOW_TF32 = True
 PIN_MEMORY = True if torch.cuda.is_available() else False
 
-from ..utils.plots                    import plotSnapshot, plotModalErrorBars, plotTimeSeries
+
+from .                                import GAVI
 
 from .pipeline                        import Pipeline, ClusteredPipeline
 from .utils                           import Dataset, MinMaxScaler, select_device, betaLinearScheduler, create_results_folder, set_seed
+from .utils                           import tanh, relu, elu, sigmoid, leakyRelu, silu
 
 from .optimizer                       import OptunaOptimizer
 
@@ -44,14 +46,6 @@ from .architectures.pinn              import PINN, BurgersPINN, Euler2DPINN, Nav
 from .architectures.shred             import SHRED
 from .architectures.binary_classifier import BinaryClassifier
 
-from .                                import GAVI
-
-# Wrapper of the activation functions
-def tanh():      return nn.Tanh()
-def relu():      return nn.ReLU()
-def elu():       return nn.ELU()
-def sigmoid():   return nn.Sigmoid()
-def leakyRelu(): return nn.LeakyReLU()
-def silu():      return nn.SiLU()
+from ..utils.plots                    import plotSnapshot, plotModalErrorBars, plotTimeSeries
 
 del os, torch, utils
